+++ src/test/kotlin/io/toasting/domain/post/application/PostServiceTest.kt
@@ -11,7 +11,7 @@ import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
 import io.kotest.matchers.string.shouldContain
 import io.mockk.every
-import io.toasting.creator.member.PostCreator
+import io.toasting.creator.post.PostCreator
 import io.toasting.domain.member.entity.Member
 import io.toasting.domain.member.exception.MemberExceptionHandler
 import io.toasting.domain.member.repository.MemberRepository
